requirements.txt
datetime
firebase_admin
flask
functions-framework
geoip2
google-api-core
google-cloud-firestore
google-cloud-bigquery
google-cloud-logging
ipaddress
python-dotenv
uuid
user_agents